Add files via upload
Task-2:-In this regression task we will predict the percentage of marks that a student is expected to score based upon the number of hours they studied. This is a simple linear regression task as it involves just two variables.
Task-5:-Performed ‘Exploratory Data Analysis’ on the provided dataset ‘SampleSuperstore’  interested in finding out the weak areas where you can work to make more profit.